requirements.txt
django==3.2.18
djangorestframework==3.14.0
drf-yasg